requirements.txt
llama-index-core
llama-index-llms-openai
llama-index-embeddings-openai
python-dotenv
pypdf
openai
llama-index-readers-file
